test: test를 하면서 두 batch를 한 클레스에 밀어넣다보니 의존관계가 복잡해지고 테스트에 어려움을 느껴 분리합니다. (#166)
